Add files via upload
Working version.
Start running the sketch and you will see the mouse trajectory being recorded on the window.
Top left is the X and Y coordinates together with the UNIX style timestamp. At the bottom left is the number of recorded positions.
Press 'q' to save a timestamped .jpg snapshot and a .txt file.
You may want to adjust the scale WinScale variable.